+++ apps/src/sites/studio/pages/pd/application/teacher_application/new.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import TeacherApplication from '@cdo/apps/code-studio/pd/application/teacher/TeacherApplicationFunc';
+import TeacherApplication from '@cdo/apps/code-studio/pd/application/teacher/TeacherApplication';
 import getScriptData from '@cdo/apps/util/getScriptData';
 
 document.addEventListener('DOMContentLoaded', function(event) {
